refactor: Tidy up product code
